+++ src/components/film-card.js
@@ -30,4 +30,9 @@ export default class Card extends AbstractComponent {
   getTemplate() {
     return getFilmCardsTemplate(this._card);
   }
+  setShowPopupHandler(handler) {
+    this._element.querySelector(`.film-card__title`).addEventListener(`click`, handler);
+    this._element.querySelector(`.film-card__comments`).addEventListener(`click`, handler);
+    this._element.querySelector(`.film-card__poster`).addEventListener(`click`, handler);
+  }
 }
