+++ src/components/show-more-button.js
@@ -1,4 +1,3 @@
-
 import AbstractComponent from "./abstract-component";
 
 export const getButtonShowMoreTemplate = () => `<button class="films-list__show-more">Show more</button>`;
@@ -6,5 +5,8 @@ export const getButtonShowMoreTemplate = () => `<button class="films-list__show-
 export default class ShowMoreButton extends AbstractComponent {
   getTemplate() {
     return getButtonShowMoreTemplate();
+  }
+  setShowMoreButtonClickHandler(handler) {
+    this._element.addEventListener(`click`, handler);
   }
 }
